business loan

+++ src/main/java/Quadraticfinal/Quadratic.java
@@ -86,13 +86,14 @@ public void isImaginary(){
 }
 }
 public double getFirstRoot(){
-   double result = (int)(-b + java.lang.Math.sqrt(getDiscriminant())) / (2*a);
+   double result = (-b + java.lang.Math.sqrt(getDiscriminant())) / (2*a);
   
    return result;
 }
 
+
 public double getSecondRoot(){
-	double result = (int) (-b - java.lang.Math.sqrt(getDiscriminant())) / (2*a);
+	double result =  (-b - java.lang.Math.sqrt(getDiscriminant())) / (2*a);
         
 	return result;
 }
